fix(ui): preserve live agent run content
